[IMP] account,*: merge audit trail
*: account_audit_trail, l10n_dk, l10n_dk_audit_trail, l10n_in, base

This commit merges account_audit_trail to account and adds improvements
on top of it:

- improved the empty screen of audit trail menu message
- added some additional filters in search for audit trail view

task-id: 3952457
