+++ addons/account/models/ir_ui_menu.py
@@ -7,5 +7,5 @@ class IrUiMenu(models.Model):
     def _load_menus_blacklist(self):
         res = super()._load_menus_blacklist()
         if not any(company.check_account_audit_trail for company in self.env.user.company_ids):
-            res.append(self.env.ref('account_audit_trail.account_audit_trail_menu').id)
+            res.append(self.env.ref('account.account_audit_trail_menu').id)
         return res
