+++ addons/account/views/mail_message_views.xml
@@ -26,10 +26,14 @@
                 <field name="account_audit_log_move_id"/>
                 <field name="author_id"/>
                 <field name="date" string="Date"/>
-                <filter string="Update Only" name="filter_tracking_value_ids"
+                <filter string="Updated Entries" name="filter_tracking_value_ids"
                         domain="[('tracking_value_ids', '!=', False)]" groups="base.group_system"/>
+                <filter string="Account Changed"
+                        name="filter_tracking_value_ids_account"
+                        domain="[('tracking_value_ids.field_id.name', '=', 'account_id')]"
+                        groups="base.group_system"/>
                 <group expand="0" string="Group By">
-                    <filter string="date" name="group_by_date" domain="[]" context="{'group_by': 'date'}"/>
+                    <filter string="Date" name="group_by_date" domain="[]" context="{'group_by': 'date'}"/>
                     <filter string="Journal Entry" name="group_by_res_id" domain="[]" context="{'group_by': 'res_id'}"/>
                 </group>
             </search>
@@ -42,9 +46,9 @@
         <field name="view_id" ref="view_message_tree_audit_log"/>
         <field name="view_mode">tree</field>
         <field name="help" type="html">
-            <p class="o_view_nocontent_smiling_face">
-            To see the audit log you have to enable the audit trail option from settings
-            </p>
+            <p class="o_view_nocontent_smiling_face">This is the Audit Trail Report</p>
+            <div>Make sure you first activate the audit trail in the accounting settings</div>
+            <a class="btn-link" type="action" name="%(action_account_config)d">Go to the accounting settings</a>
         </field>
         <field name="domain">[
             ('model', '=', 'account.move'),
