+++ addons/account/views/res_config_settings_views.xml
@@ -385,6 +385,11 @@
                                 <field name="quick_edit_mode" placeholder="Disabled"/>
                             </setting>
                         </block>
+                        <block title="Audit Trail" name="audit_trail_setting_container">
+                            <setting string="Audit Trail" company_dependent="1" help="Activate Audit Trail">
+                                <field name="check_account_audit_trail"/>
+                            </setting>
+                        </block>
                     </app>
                 </xpath>
             </field>
