--- addons/account_audit_trail/i18n/account_audit_trail.pot
@@ -1,138 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-# 	* account_audit_trail
-#
-msgid ""
-msgstr ""
-"Project-Id-Version: Odoo Server saas~17.3\n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2024-06-03 08:43+0000\n"
-"PO-Revision-Date: 2024-06-03 08:43+0000\n"
-"Last-Translator: \n"
-"Language-Team: \n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: \n"
-"Plural-Forms: \n"
-
-#. module: account_audit_trail
-#: model_terms:ir.ui.view,arch_db:account_audit_trail.res_config_settings_view_form_inherit_account_audit_trail
-msgid "Activate Audit Trail"
-msgstr ""
-
-#. module: account_audit_trail
-#: model:ir.actions.act_window,name:account_audit_trail.action_account_audit_trail_report
-#: model:ir.model.fields,field_description:account_audit_trail.field_res_company__check_account_audit_trail
-#: model:ir.model.fields,field_description:account_audit_trail.field_res_config_settings__check_account_audit_trail
-#: model:ir.ui.menu,name:account_audit_trail.account_audit_trail_menu
-#: model_terms:ir.ui.view,arch_db:account_audit_trail.res_config_settings_view_form_inherit_account_audit_trail
-msgid "Audit Trail"
-msgstr ""
-
-#. module: account_audit_trail
-#. odoo-python
-#: code:addons/account_audit_trail/models/res_config_settings.py:0
-msgid "Can't disable audit trail when there are existing records."
-msgstr ""
-
-#. module: account_audit_trail
-#. odoo-python
-#: code:addons/account_audit_trail/models/mail_message.py:0
-msgid "Changed"
-msgstr ""
-
-#. module: account_audit_trail
-#: model:ir.model,name:account_audit_trail.model_res_company
-msgid "Companies"
-msgstr ""
-
-#. module: account_audit_trail
-#: model:ir.model,name:account_audit_trail.model_res_config_settings
-msgid "Config Settings"
-msgstr ""
-
-#. module: account_audit_trail
-#: model_terms:ir.ui.view,arch_db:account_audit_trail.view_message_tree_audit_log_search
-msgid "Date"
-msgstr ""
-
-#. module: account_audit_trail
-#: model:ir.model.fields,field_description:account_audit_trail.field_mail_mail__account_audit_log_preview
-#: model:ir.model.fields,field_description:account_audit_trail.field_mail_message__account_audit_log_preview
-msgid "Description"
-msgstr ""
-
-#. module: account_audit_trail
-#: model_terms:ir.ui.view,arch_db:account_audit_trail.res_config_settings_view_form_inherit_account_audit_trail
-msgid "Go to Audit Trail"
-msgstr ""
-
-#. module: account_audit_trail
-#: model_terms:ir.ui.view,arch_db:account_audit_trail.view_message_tree_audit_log_search
-msgid "Group By"
-msgstr ""
-
-#. module: account_audit_trail
-#: model:ir.model,name:account_audit_trail.model_account_move
-#: model:ir.model.fields,field_description:account_audit_trail.field_mail_mail__account_audit_log_move_id
-#: model:ir.model.fields,field_description:account_audit_trail.field_mail_message__account_audit_log_move_id
-#: model_terms:ir.ui.view,arch_db:account_audit_trail.view_message_tree_audit_log_search
-msgid "Journal Entry"
-msgstr ""
-
-#. module: account_audit_trail
-#: model:ir.model,name:account_audit_trail.model_ir_ui_menu
-msgid "Menu"
-msgstr ""
-
-#. module: account_audit_trail
-#: model:ir.model,name:account_audit_trail.model_mail_message
-msgid "Message"
-msgstr ""
-
-#. module: account_audit_trail
-#: model_terms:ir.ui.view,arch_db:account_audit_trail.view_message_tree_audit_log_search
-msgid "Messages Search"
-msgstr ""
-
-#. module: account_audit_trail
-#. odoo-python
-#: code:addons/account_audit_trail/models/mail_message.py:0
-msgid "Operation not supported"
-msgstr ""
-
-#. module: account_audit_trail
-#: model:ir.model.fields,field_description:account_audit_trail.field_mail_mail__show_audit_log
-#: model:ir.model.fields,field_description:account_audit_trail.field_mail_message__show_audit_log
-msgid "Show Audit Log"
-msgstr ""
-
-#. module: account_audit_trail
-#. odoo-python
-#: code:addons/account_audit_trail/models/account_move.py:0
-msgid ""
-"To keep the audit trail, you can not delete journal entries once they have been posted.\n"
-"Instead, you can cancel the journal entry."
-msgstr ""
-
-#. module: account_audit_trail
-#: model_terms:ir.actions.act_window,help:account_audit_trail.action_account_audit_trail_report
-msgid ""
-"To see the audit log you have to enable the audit trail option from settings"
-msgstr ""
-
-#. module: account_audit_trail
-#: model_terms:ir.ui.view,arch_db:account_audit_trail.view_message_tree_audit_log_search
-msgid "Update Only"
-msgstr ""
-
-#. module: account_audit_trail
-#. odoo-python
-#: code:addons/account_audit_trail/models/mail_message.py:0
-msgid "Updated"
-msgstr ""
-
-#. module: account_audit_trail
-#: model_terms:ir.ui.view,arch_db:account_audit_trail.view_message_tree_audit_log_search
-msgid "date"
-msgstr ""
