--- addons/account_audit_trail/i18n/ru.po
@@ -1,141 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-# 	* account_audit_trail
-#
-# Translators:
-msgid ""
-msgstr ""
-"Project-Id-Version: Odoo Server 17.0\n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2024-02-12 10:36+0000\n"
-"PO-Revision-Date: 2024-01-30 15:14+0400\n"
-"Last-Translator: \n"
-"Language-Team: Russian (https://app.transifex.com/odoo/teams/41243/ru/)\n"
-"Language: ru\n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: \n"
-"Plural-Forms: nplurals=4; plural=(n%10==1 && n%100!=11 ? 0 : n%10>=2 && n%10<=4 && (n%100<12 || n%100>14) ? 1 : n%10==0 || (n%10>=5 && n%10<=9) || (n%100>=11 && n%100<=14)? 2 : 3);\n"
-
-#. module: account_audit_trail
-#: model_terms:ir.ui.view,arch_db:account_audit_trail.res_config_settings_view_form_inherit_account_audit_trail
-msgid "Activate Audit Trail"
-msgstr "Активировать журнал аудита"
-
-#. module: account_audit_trail
-#: model:ir.actions.act_window,name:account_audit_trail.action_account_audit_trail_report
-#: model:ir.model.fields,field_description:account_audit_trail.field_res_company__check_account_audit_trail
-#: model:ir.model.fields,field_description:account_audit_trail.field_res_config_settings__check_account_audit_trail
-#: model:ir.ui.menu,name:account_audit_trail.account_audit_trail_menu
-#: model_terms:ir.ui.view,arch_db:account_audit_trail.res_config_settings_view_form_inherit_account_audit_trail
-msgid "Audit Trail"
-msgstr "Журнал аудита"
-
-#. module: account_audit_trail
-#. odoo-python
-#: code:addons/account_audit_trail/models/res_config_settings.py:0
-msgid "Can't disable audit trail when there are existing records."
-msgstr "Невозможно отключить журнал аудита при наличии существующих записей."
-
-#. module: account_audit_trail
-#. odoo-python
-#: code:addons/account_audit_trail/models/mail_message.py:0
-msgid "Changed"
-msgstr "Изменено"
-
-#. module: account_audit_trail
-#: model:ir.model,name:account_audit_trail.model_res_company
-msgid "Companies"
-msgstr "Компании"
-
-#. module: account_audit_trail
-#: model:ir.model,name:account_audit_trail.model_res_config_settings
-msgid "Config Settings"
-msgstr "Параметры конфигурации"
-
-#. module: account_audit_trail
-#: model_terms:ir.ui.view,arch_db:account_audit_trail.view_message_tree_audit_log_search
-msgid "Date"
-msgstr "Дата"
-
-#. module: account_audit_trail
-#: model:ir.model.fields,field_description:account_audit_trail.field_mail_mail__account_audit_log_preview
-#: model:ir.model.fields,field_description:account_audit_trail.field_mail_message__account_audit_log_preview
-msgid "Description"
-msgstr "Описание"
-
-#. module: account_audit_trail
-#: model_terms:ir.ui.view,arch_db:account_audit_trail.res_config_settings_view_form_inherit_account_audit_trail
-msgid "Go to Audit Trail"
-msgstr "Перейти к журналу аудита"
-
-#. module: account_audit_trail
-#: model_terms:ir.ui.view,arch_db:account_audit_trail.view_message_tree_audit_log_search
-msgid "Group By"
-msgstr "Группировать по"
-
-#. module: account_audit_trail
-#: model:ir.model,name:account_audit_trail.model_account_move
-#: model:ir.model.fields,field_description:account_audit_trail.field_mail_mail__account_audit_log_move_id
-#: model:ir.model.fields,field_description:account_audit_trail.field_mail_message__account_audit_log_move_id
-#: model_terms:ir.ui.view,arch_db:account_audit_trail.view_message_tree_audit_log_search
-msgid "Journal Entry"
-msgstr "Запись в журнале"
-
-#. module: account_audit_trail
-#: model:ir.model,name:account_audit_trail.model_ir_ui_menu
-msgid "Menu"
-msgstr "Меню"
-
-#. module: account_audit_trail
-#: model:ir.model,name:account_audit_trail.model_mail_message
-msgid "Message"
-msgstr "Сообщение"
-
-#. module: account_audit_trail
-#: model_terms:ir.ui.view,arch_db:account_audit_trail.view_message_tree_audit_log_search
-msgid "Messages Search"
-msgstr "Поиск сообщений"
-
-#. module: account_audit_trail
-#. odoo-python
-#: code:addons/account_audit_trail/models/mail_message.py:0
-msgid "Operation not supported"
-msgstr "Операция не поддерживается"
-
-#. module: account_audit_trail
-#: model:ir.model.fields,field_description:account_audit_trail.field_mail_mail__show_audit_log
-#: model:ir.model.fields,field_description:account_audit_trail.field_mail_message__show_audit_log
-msgid "Show Audit Log"
-msgstr "Показать журнал аудита"
-
-#. module: account_audit_trail
-#. odoo-python
-#: code:addons/account_audit_trail/models/account_move.py:0
-msgid ""
-"To keep the audit trail, you can not delete journal entries once they have been posted.\n"
-"Instead, you can cancel the journal entry."
-msgstr ""
-"Чтобы сохранить аудиторский след, вы не можете удалять записи журнала после того, как они были опубликованы.\n"
-"Вместо этого можно отменить запись в журнале."
-
-#. module: account_audit_trail
-#: model_terms:ir.actions.act_window,help:account_audit_trail.action_account_audit_trail_report
-msgid "To see the audit log you have to enable the audit trail option from settings"
-msgstr "Чтобы увидеть журнал аудита, необходимо включить опцию аудиторского следа в настройках"
-
-#. module: account_audit_trail
-#: model_terms:ir.ui.view,arch_db:account_audit_trail.view_message_tree_audit_log_search
-msgid "Update Only"
-msgstr "Только обновление"
-
-#. module: account_audit_trail
-#. odoo-python
-#: code:addons/account_audit_trail/models/mail_message.py:0
-msgid "Updated"
-msgstr "Обновлено"
-
-#. module: account_audit_trail
-#: model_terms:ir.ui.view,arch_db:account_audit_trail.view_message_tree_audit_log_search
-msgid "date"
-msgstr "дата"
