--- addons/account_audit_trail/models/__init__.py
@@ -1,8 +1,0 @@
-# -*- coding: utf-8 -*-
-# Part of Odoo. See LICENSE file for full copyright and licensing details.
-
-from . import account_move
-from . import ir_ui_menu
-from . import mail_message
-from . import res_company
-from . import res_config_settings
