--- addons/l10n_de_audit_trail/__manifest__.py
@@ -1,15 +1,0 @@
-{
-    'name': 'Germany - audit trail',
-    'version': '1.0',
-    'description': """
-This module is a bridge to auto-install Audit Trail with Germany
-    """,
-    'summary': "Audit trail",
-    'depends': [
-        'l10n_de',
-        'account_audit_trail'
-    ],
-    'installable': True,
-    'auto_install': ['l10n_de'],
-    'license': 'LGPL-3',
-}
