+++ addons/l10n_dk/__manifest__.py
@@ -92,7 +92,6 @@ Produkt setup:
         'base_iban',
         'base_vat',
         'account',
-        'account_audit_trail',
     ],
     'auto_install': ['account'],
     'data': [
